requirements.txt
requests>=2.28.0
pytest>=7.0.0
pytest-html>=3.2.0
pytest-xdist>=3.0.0
pytest-rerunfailures>=10.0
pytest-timeout>=2.1.0
allure-pytest>=2.9.0
PyYAML>=6.0
pandas>=1.5.0
openpyxl>=3.0.0
jsonschema>=4.17.0
deepdiff>=6.2.0
pymysql>=1.0.0
python-dotenv>=0.21.0